relab_backend/requirements.txt
# requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
django-admin-interface==0.28.8
mercadopago==2.2.0